[expand] applied review feedback
