+++ server/diagram-api/src/main/java/io/typefox/sprotty/api/IDiagramExpansionListener.java
@@ -1,3 +1,9 @@
+/*
+ * Copyright (C) 2017 TypeFox and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ */
 package io.typefox.sprotty.api;
 
 /**
